주문 취소 로직 업데이트했습니다. (보안 강화, 로직 수정)

+++ src/main/java/com/example/janghj/service/OrderService.java
@@ -4,7 +4,6 @@ import com.example.janghj.config.security.UserDetailsImpl;
 import com.example.janghj.domain.Delivery;
 import com.example.janghj.domain.Order;
 import com.example.janghj.domain.Product.Product;
-import com.example.janghj.domain.User.TypeOfCash;
 import com.example.janghj.domain.User.User;
 import com.example.janghj.repository.OrderRepository;
 import com.example.janghj.repository.ProductRepository;
@@ -13,7 +12,9 @@ import com.example.janghj.web.dto.OrderProduct;
 import com.example.janghj.web.dto.OrderWebDto;
 import lombok.RequiredArgsConstructor;
 import lombok.SneakyThrows;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -24,14 +25,14 @@ public class OrderService {
     private final OrderRepository orderRepository;
     private final ProductRepository productRepository;
     private final UserRepository userRepository;
-    private final UserService userService;
 
+    @Transactional(readOnly = true, rollbackFor = Throwable.class)
     public Order order(UserDetailsImpl nowUser, OrderWebDto orderWebDto) {
-//        User user = userRepository.findById(nowUser.getId()).orElseThrow(
-//                () -> new NullPointerException("해당 유저가 존재하지 않습니다. id  = " + nowUser.getId()));
+        User user = userRepository.findById(nowUser.getId()).orElseThrow(
+                () -> new NullPointerException("해당 유저가 존재하지 않습니다. id  = " + nowUser.getId()));
 
-        User user = userRepository.findById(1L).orElseThrow(
-                () -> new NullPointerException("해당 유저가 존재하지 않습니다. id  = " + 1L));
+//        User user = userRepository.findById(1L).orElseThrow(
+//                () -> new NullPointerException("해당 유저가 존재하지 않습니다. id  = " + 1L));
 
         Order order = new Order(user);
 
@@ -48,35 +49,59 @@ public class OrderService {
     }
 
     @SneakyThrows // try-catch 기능을 대체해줌
+    @Transactional(readOnly = true, rollbackFor = Throwable.class)
     public OrderProduct createOrder(Long productId, int quantity, Order order) {
         Product product = (Product) productRepository.findById(productId).orElseThrow(
                 () -> new NullPointerException("해당 상품이 없습니다. productId =" + productId));
+
         int amount = product.getPrice() * quantity;
         return new OrderProduct(product, product.getPrice(), order, quantity, amount);
     }
 
+    @Transactional(readOnly = true, rollbackFor = Throwable.class)
     public void payForTheOrder(UserDetailsImpl nowUser, Long orderId) {
-        Order order = getOrder(orderId);
+        findByOrder(orderId);
     }
 
+    @Transactional(readOnly = true, rollbackFor = Throwable.class)
     public void payForTheOrders(UserDetailsImpl nowUser) {
-        List<Order> orders = getOrders(nowUser);
+
     }
 
+    @Transactional(readOnly = true, rollbackFor = Throwable.class)
+    public void deliveryStart(UserDetailsImpl nowUser) {
 
-    public Order getOrder(Long orderId) {
+    }
+
+    @Transactional(readOnly = true, rollbackFor = Throwable.class)
+    public void deliveryArrive(UserDetailsImpl nowUser) {
+
+    }
+
+    @Transactional(readOnly = true, rollbackFor = Throwable.class)
+    public Order findByOrder(Long orderId) {
         Order order = orderRepository.findById(orderId).orElseThrow(
                 () -> new NullPointerException("해당 주문이 존재하지 않습니다. itemId = " + orderId)
         );
         return order;
     }
 
-    public List<Order> getOrders(UserDetailsImpl nowUser) {
+    public List<Order> findByOrders(UserDetailsImpl nowUser) {
         return orderRepository.findAllByUserId(nowUser.getId());
     }
 
-    public void orderCancel(Long orderId) {
+    @Transactional(readOnly = true, rollbackFor = Throwable.class)
+    public void orderCancel(UserDetailsImpl nowUser, Long orderId) {
+        Order deleteOrder = orderRepository.findById(orderId).orElseThrow(
+                () -> new NullPointerException("해당 주문이 존재하지 않습니다. itemId = " + orderId)
+        );
+
+        if (!deleteOrder.getUser().equals(nowUser.getUser())) {
+            throw new AccessDeniedException("유저(" + nowUser.getId() + ") 가 다른 유저(" + deleteOrder.getUser().getId() + ")에 접근하여 주문을 삭제하려고 합니다.");
+        }
+
         orderRepository.deleteById(orderId);
+
     }
 
 }
